refactor(Components): WIP towards refactoring Payment Form to component
